% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{DJTmps}
\alias{DJTmps}
\title{December and January Temperature Estimation}
\usage{
DJTmps(dSlice)
}
\arguments{
\item{dSlice}{TODO}
}
\description{
(Model) Estimate Dec Temp in Year-1 from Jan in Year 
and Jan Temp in Y+1 from Dec in Y
}
