% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindArea}
\alias{FindArea}
\title{Find Area}
\usage{
FindArea(area, zmn, zmx, ztarget)
}
\arguments{
\item{area}{TODO}

\item{zmn}{TODO}

\item{zmx}{TODO}

\item{ztarget}{TODO}
}
\description{
(Model and Report) Find the area between two depths in a lake
}

