% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindDay}
\alias{FindDay}
\title{Find Day}
\usage{
FindDay(endtemp, startday, starttemp, tslope)
}
\arguments{
\item{endtemp}{TODO}

\item{startday}{TODO}

\item{starttemp}{TODO}

\item{tslope}{TODO}
}
\description{
(Report) Calculate a julian day for a given temperature based on a known
day with known temperature and known slope of temperature change
}

