% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindDepth}
\alias{FindDepth}
\title{Find Depth}
\usage{
FindDepth(zmx, tx, tn, js, jm, je, zm, zj, sp, jd, targettemp)
}
\arguments{
\item{zmx}{TODO}

\item{tx}{TODO}

\item{tn}{TODO}

\item{js}{TODO}

\item{jm}{TODO}

\item{je}{TODO}

\item{zm}{TODO}

\item{zj}{TODO}

\item{sp}{TODO}

\item{jd}{TODO}

\item{targettemp}{TODO}
}
\description{
(Report) Find the depth at a particular temperature in a lake
}
