% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindHabitat}
\alias{FindHabitat}
\title{Find Habitat Volume and Area}
\usage{
FindHabitat(tx, tn, js, jm, je, zm, zj, sp, jd, zmx, area, zmn, lakevol,
  upperlimit, lowerlimit)
}
\arguments{
\item{tx}{TODO}

\item{tn}{TODO}

\item{js}{TODO}

\item{jm}{TODO}

\item{je}{TODO}

\item{zm}{TODO}

\item{zj}{TODO}

\item{sp}{TODO}

\item{jd}{TODO}

\item{zmx}{TODO}

\item{area}{TODO}

\item{zmn}{TODO}

\item{lakevol}{TODO}

\item{upperlimit}{TODO}

\item{lowerlimit}{TODO}
}
\description{
(Report) Find the volume and area between two temperatures in a lake
}
