% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindHabZ}
\alias{FindHabZ}
\title{Find Habitat Depth}
\usage{
FindHabZ(tx, tn, js, jm, je, zm, zj, sp, jd, zmx, area, zmn, T1, T2)
}
\arguments{
\item{tx}{TODO}

\item{tn}{TODO}

\item{js}{TODO}

\item{jm}{TODO}

\item{je}{TODO}

\item{zm}{TODO}

\item{zj}{TODO}

\item{sp}{TODO}

\item{jd}{TODO}

\item{zmx}{TODO}

\item{area}{TODO}

\item{zmn}{TODO}

\item{T1}{TODO}

\item{T2}{TODO}
}
\description{
(Report) Find the depth for a given temperature in a lake
}

