% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindTemp}
\alias{FindTemp}
\title{Find Temperature}
\usage{
FindTemp(endday, startday, starttemp, tslope)
}
\arguments{
\item{endday}{TODO}

\item{startday}{TODO}

\item{starttemp}{TODO}

\item{tslope}{TODO}
}
\description{
(Report) Calculate a temperature for a given day based on a known day
with known temperature and known slope of temperature change
}
