% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{FindVol}
\alias{FindVol}
\title{Find Volume}
\usage{
FindVol(area, zmn, zmx, zstart, zend)
}
\arguments{
\item{area}{TODO}

\item{zmn}{TODO}

\item{zmx}{TODO}

\item{zstart}{TODO}

\item{zend}{TODO}
}
\description{
(Report) Find the volume between two depths in a lake
}

