% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_thermopic_data}
\alias{get_thermopic_data}
\title{Get data from a thermopic project}
\usage{
get_thermopic_data(data, path = NULL, in_or_out = c("DataIn", "DataOut"))
}
\arguments{
\item{data}{Character (or factor) containing a path to a CSV, or a data frame}

\item{path}{Thermopic project root path}

\item{in_or_out}{Either DataIn or DataOut}
}
\description{
Get data from a thermopic project
}

