% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{print_directory_tree}
\alias{print_directory_tree}
\title{Print directory tree}
\usage{
print_directory_tree(root)
}
\arguments{
\item{root}{directory to depict visually}
}
\description{
Print directory tree
}
