% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{SpStats}
\alias{SpStats}
\title{Spatial Habitat Statistics}
\usage{
SpStats(hssub, JM)
}
\arguments{
\item{hssub}{TODO}

\item{JM}{TODO}
}
\description{
(Report) Calculate Annual Summary Habitat Space Statistics
based on Daily Predictions from the STM model for the period
Spring isothermal at 4C to Fall/Autumn isothermal at 4C
Detects Presence of One Continuous  or Two Seasons of Availability
For Stratified and Unstratified Lakes Respectively
}
