% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{stm}
\alias{stm}
\title{Seasonal Temperature Model}
\usage{
stm(tx, tn, js, jm, je, zm, zj, sp, jd, z)
}
\arguments{
\item{tx}{STM parameter maximum surface temperature on JM}

\item{tn}{STM parameter hypolimnetic temperature at onset of stratification JS}

\item{js}{STM parameter day of onset of stratification}

\item{jm}{STM parameter day of peak surface temperature}

\item{je}{STM parameter day of end of stratification}

\item{zm}{STM parameter maximum thermocline depth}

\item{zj}{STM parameter number of days after JS when ZTH attains ZM/2}

\item{sp}{STM parameter steepness of transition from epi- to hypo-limnion}

\item{jd}{FIXME: Julian day?}

\item{z}{FIXME: depth?}
}
\description{
(Model and Report) Find the water temperature at a given depth on a given day in a lake
}
