% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{SunPosition}
\alias{SunPosition}
\title{Get Solar Elevation and Azimuth}
\usage{
SunPosition(year, month, day, hour = 12, min = 0, sec = 0, lat, long)
}
\arguments{
\item{year}{TODO}

\item{month}{TODO}

\item{day}{TODO}

\item{hour}{TODO}

\item{min}{TODO}

\item{sec}{TODO}

\item{lat}{TODO}

\item{long}{TODO}
}
\description{
(Model) Get solar elevation and azimuth for a specified date and location.
The implementation follows the Astronomer's Almanac algorithm.
}
\references{
Joseph J. Michalsky. 
The astronomical almanacs algorithm for approximate solar position (19502050). 
Solar Energy, 40(3):227235, 1988.
It's straightforward to translate it in any other language:
http://stackoverflow.com/questions/257717/position-of-the-sun-given-time-of-day-and-lat-long

http://www.esrl.noaa.gov/gmd/grad/solcalc/
Checked against NOAA on-line calculator
}
