% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_routines.R
\name{tdatev}
\alias{tdatev}
\title{Interpolate Julian Dates}
\usage{
tdatev(tvec, tref, tdir, yy)
}
\arguments{
\item{tvec}{data frame with mean monthly temperatures from Dec (Y-1) 
through Jan-Dec (Y) to Jan(Y+1)}

\item{tref}{threshold temperature C}

\item{tdir}{value indicates whether the temperature is rising (1) or falling (-1)}

\item{yy}{calendar year (allows day counts to be adjusted for leap years)}
}
\description{
(Model) Interpolate Julian dates for reference temperatures 
in spring and fall from monthly mean temperatures.
}
