% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{temporary_thermopic_directory}
\alias{temporary_thermopic_directory}
\title{Temporary thermopic directory}
\usage{
temporary_thermopic_directory()
}
\description{
Temporary thermopic directory
}
