% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermopic.R
\docType{package}
\name{thermopic}
\alias{thermopic}
\alias{thermopic-package}
\title{thermopic: Predicting and visualizing the seasonal available of thermal habitat in lakes}
\description{
A ThermoPic is a picture of thermal habitat in a lake. This software serves
three purposes. First, it predicts the seasonal temperature cycle in lakes
based on easily measured variables. Second, it calculates and reports thermal
habitat statistics. Third, it generates a ThermoPic plot. The model used to 
predict temperature is referred to as the Seasonal Temperature Model (STM). 
It was created by Minns et al. (2016) based on data from lakes in the province 
of Ontario (Canada).
}
\details{
The thermopic package works with thermopic projects, which are filesystem directories
with a special structure

TODO: describe how to use the package with reference to the original pdf document
}
