% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{thermopic_model}
\alias{thermopic_model}
\title{Fit Thermopic Model}
\usage{
thermopic_model(path, laked, sited, lakep_output_file = "3_Model_Inputs.csv",
  STM_output_file = "4_STM_Parameters.csv",
  sited_output_file = "tmp_ClimMetrics.csv",
  siteo_output_file = "tmp_IceClimMetrics.csv", year_fix = 2010)
}
\arguments{
\item{path}{Character string containing the path to the root of a thermopic project}

\item{laked}{A \code{\link{data.frame}} or \code{csv} file with schema defined in section C4 of the \href{../doc/ThermoPic_Guide_v3b.pdf}{ThermoPic Guide} under the \code{1_Lake.csv} sub-heading. If a \code{csv} file is used, please place it within the \code{DataIn} folder under the \pkg{thermopic} project in \code{path}.}

\item{sited}{A \code{\link{data.frame}} or \code{csv} file with schema defined in section C4 of the \href{../doc/ThermoPic_Guide_v3b.pdf}{ThermoPic Guide} under the \code{2_Climate.csv} sub-heading. If a \code{csv} file is used, please place it within the \code{DataIn} folder under the \pkg{thermopic} project in \code{path}.}

\item{lakep_output_file}{The name of an output \code{csv} file with schema defined in section C5 of the \href{../doc/ThermoPic_Guide_v3b.pdf}{ThermoPic Guide} under the \code{3_Model_Inputs.csv} sub-heading.}

\item{STM_output_file}{The name of an output \code{csv} file with schema defined in section C5 of the \href{../doc/ThermoPic_Guide_v3b.pdf}{ThermoPic Guide} under the \code{4_STM_Parameters.csv} sub-heading.}

\item{sited_output_file}{The name of an output temp file used mainly for debugging}

\item{siteo_output_file}{The name of an output temp file used mainly for debugging}

\item{year_fix}{Integer giving a calender year for simplifing the code by
assigning "year_fix" for calculating Day of Year and Solar elevation/angle.
Exact year not needed because because program works with Temperature norms
for Specified Period. Solar angle for a given day of year varies on 4 year
cycle, but variation is very small.}
}
\value{
s3 object of class thermopic_model, containing output dataframes
}
\description{
Fit Thermopic Model
}
