% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{thermopic_model}
\alias{thermopic_model}
\title{Fit Thermopic Model}
\usage{
thermopic_model(path, laked, sited, UserOptions, year_fix = 2010)
}
\arguments{
\item{path}{Character string containing the path to the root of a thermopic project}

\item{laked}{TODO}

\item{sited}{TODO}

\item{UserOptions}{TODO}

\item{year_fix}{TODO}
}
\value{
s3 object of class thermopic_model, containing output dataframes
}
\description{
Fit Thermopic Model
}

