% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{thermopic_model}
\alias{thermopic_model}
\title{Fit Thermopic Model}
\usage{
thermopic_model(path, laked, sited, sited_output_file = "tmp_ClimMetrics.csv",
  lakep_output_file = "3_Model_Inputs.csv",
  siteo_output_file = "tmp_IceClimMetrics.csv",
  STM_output_file = "4_STM_Parameters.csv", year_fix = 2010)
}
\arguments{
\item{path}{Character string containing the path to the root of a thermopic project}

\item{laked}{TODO}

\item{sited}{TODO}

\item{sited_output_file}{TODO}

\item{lakep_output_file}{TODO}

\item{siteo_output_file}{TODO}

\item{STM_output_file}{TODO}

\item{year_fix}{TODO}
}
\value{
s3 object of class thermopic_model, containing output dataframes
}
\description{
Fit Thermopic Model
}
